refactor find_includes
