+++ legacy/builder/create_cmake_rule.go
@@ -26,11 +26,11 @@ import (
 	properties "github.com/arduino/go-properties-orderedmap"
 	"golang.org/x/exp/slices"
 
+	"github.com/arduino/arduino-cli/arduino/builder/utils"
 	"github.com/arduino/arduino-cli/arduino/globals"
 	"github.com/arduino/arduino-cli/legacy/builder/builder_utils"
 	"github.com/arduino/arduino-cli/legacy/builder/constants"
 	"github.com/arduino/arduino-cli/legacy/builder/types"
-	"github.com/arduino/arduino-cli/legacy/builder/utils"
 )
 
 type ExportProjectCMake struct {
