+++ legacy/builder/phases/libraries_builder.go
@@ -18,12 +18,12 @@ package phases
 import (
 	"strings"
 
+	"github.com/arduino/arduino-cli/arduino/builder/utils"
 	"github.com/arduino/arduino-cli/arduino/libraries"
 	f "github.com/arduino/arduino-cli/internal/algorithms"
 	"github.com/arduino/arduino-cli/legacy/builder/builder_utils"
 	"github.com/arduino/arduino-cli/legacy/builder/constants"
 	"github.com/arduino/arduino-cli/legacy/builder/types"
-	"github.com/arduino/arduino-cli/legacy/builder/utils"
 	"github.com/arduino/go-paths-helper"
 	"github.com/arduino/go-properties-orderedmap"
 	"github.com/pkg/errors"
@@ -37,7 +37,8 @@ type LibrariesBuilder struct{}
 func (s *LibrariesBuilder) Run(ctx *types.Context) error {
 	librariesBuildPath := ctx.LibrariesBuildPath
 	buildProperties := ctx.BuildProperties
-	includes := f.Map(ctx.IncludeFolders.AsStrings(), utils.WrapWithHyphenI)
+	includesFolders := ctx.SketchLibrariesDetector.IncludeFolders()
+	includes := f.Map(includesFolders.AsStrings(), utils.WrapWithHyphenI)
 	libs := ctx.SketchLibrariesDetector.ImportedLibraries()
 
 	if err := librariesBuildPath.MkdirAll(); err != nil {
