+++ legacy/builder/phases/sketch_builder.go
@@ -16,10 +16,10 @@
 package phases
 
 import (
+	"github.com/arduino/arduino-cli/arduino/builder/utils"
 	f "github.com/arduino/arduino-cli/internal/algorithms"
 	"github.com/arduino/arduino-cli/legacy/builder/builder_utils"
 	"github.com/arduino/arduino-cli/legacy/builder/types"
-	"github.com/arduino/arduino-cli/legacy/builder/utils"
 	"github.com/pkg/errors"
 )
 
@@ -28,7 +28,8 @@ type SketchBuilder struct{}
 func (s *SketchBuilder) Run(ctx *types.Context) error {
 	sketchBuildPath := ctx.SketchBuildPath
 	buildProperties := ctx.BuildProperties
-	includes := f.Map(ctx.IncludeFolders.AsStrings(), utils.WrapWithHyphenI)
+	includesFolders := ctx.SketchLibrariesDetector.IncludeFolders()
+	includes := f.Map(includesFolders.AsStrings(), utils.WrapWithHyphenI)
 
 	if err := sketchBuildPath.MkdirAll(); err != nil {
 		return errors.WithStack(err)
