+++ legacy/builder/utils/utils.go
@@ -20,41 +20,11 @@ import (
 	"os"
 	"os/exec"
 	"strings"
-	"unicode"
 
 	f "github.com/arduino/arduino-cli/internal/algorithms"
 	"github.com/arduino/arduino-cli/legacy/builder/types"
-	paths "github.com/arduino/go-paths-helper"
 	"github.com/pkg/errors"
-	"golang.org/x/text/runes"
-	"golang.org/x/text/transform"
-	"golang.org/x/text/unicode/norm"
 )
-
-var SOURCE_CONTROL_FOLDERS = map[string]bool{"CVS": true, "RCS": true, ".git": true, ".github": true, ".svn": true, ".hg": true, ".bzr": true, ".vscode": true, ".settings": true, ".pioenvs": true, ".piolibdeps": true}
-
-// FilterOutHiddenFiles is a ReadDirFilter that exclude files with a "." prefix in their name
-var FilterOutHiddenFiles = paths.FilterOutPrefixes(".")
-
-// FilterOutSCCS is a ReadDirFilter that excludes known VSC or project files
-func FilterOutSCCS(file *paths.Path) bool {
-	return !SOURCE_CONTROL_FOLDERS[file.Base()]
-}
-
-// FilterReadableFiles is a ReadDirFilter that accepts only readable files
-func FilterReadableFiles(file *paths.Path) bool {
-	// See if the file is readable by opening it
-	f, err := file.Open()
-	if err != nil {
-		return false
-	}
-	f.Close()
-	return true
-}
-
-func WrapWithHyphenI(value string) string {
-	return "\"-I" + value + "\""
-}
 
 func printableArgument(arg string) string {
 	if strings.ContainsAny(arg, "\"\\ \t") {
@@ -126,29 +96,6 @@ func ExecCommand(ctx *types.Context, command *exec.Cmd, stdout int, stderr int) 
 	return outbytes, errbytes, errors.WithStack(err)
 }
 
-func FindFilesInFolder(dir *paths.Path, recurse bool, extensions ...string) (paths.PathList, error) {
-	fileFilter := paths.AndFilter(
-		FilterOutHiddenFiles,
-		FilterOutSCCS,
-		paths.FilterOutDirectories(),
-		FilterReadableFiles,
-	)
-	if len(extensions) > 0 {
-		fileFilter = paths.AndFilter(
-			paths.FilterSuffixes(extensions...),
-			fileFilter,
-		)
-	}
-	if recurse {
-		dirFilter := paths.AndFilter(
-			FilterOutHiddenFiles,
-			FilterOutSCCS,
-		)
-		return dir.ReadDirRecursiveFiltered(dirFilter, fileFilter)
-	}
-	return dir.ReadDir(fileFilter)
-}
-
 type loggerAction struct {
 	onlyIfVerbose bool
 	warn          bool
@@ -168,12 +115,4 @@ func (l *loggerAction) Run(ctx *types.Context) error {
 
 func LogIfVerbose(warn bool, msg string) types.Command {
 	return &loggerAction{onlyIfVerbose: true, warn: warn, msg: msg}
-}
-
-// Normalizes an UTF8 byte slice
-// TODO: use it more often troughout all the project (maybe on logger interface?)
-func NormalizeUTF8(buf []byte) []byte {
-	t := transform.Chain(norm.NFD, runes.Remove(runes.In(unicode.Mn)), norm.NFC)
-	result, _, _ := transform.Bytes(t, buf)
-	return result
 }
